login error messages + longer auth token
